+++ common/src/main/java/de/maxhenkel/voicechat/config/PlayerVolumeConfig.java
@@ -11,38 +11,61 @@ import java.util.UUID;
 public class PlayerVolumeConfig extends PropertyConfig {
 
     private final Map<UUID, Double> volumes;
+    private final Map<String, Double> categoryVolumes;
 
     public PlayerVolumeConfig(Path path) {
         super(path);
         Map<String, Object> entries = getEntries();
-        volumes = new HashMap<>(entries.size());
+        volumes = new HashMap<>();
+        categoryVolumes = new HashMap<>();
         for (Map.Entry<String, Object> entry : entries.entrySet()) {
+            double volume = Double.parseDouble(entry.getValue().toString());
             try {
-                volumes.put(UUID.fromString(entry.getKey()), Double.parseDouble(entry.getValue().toString()));
-            } catch (Exception e) {
+                volumes.put(UUID.fromString(entry.getKey()), volume);
+            } catch (IllegalArgumentException e) {
+                categoryVolumes.put(entry.getKey(), volume);
             }
         }
     }
 
-    public double getVolume(UUID uuid, double def) {
+    public double getPlayerVolume(UUID uuid, double def) {
         Double volume = volumes.get(uuid);
         if (volume == null) {
-            return setVolume(uuid, def);
+            return def;
         }
         return volume;
     }
 
-    public double getVolume(UUID playerID) {
-        return getVolume(playerID, 1D);
+    public double getPlayerVolume(UUID playerID) {
+        return getPlayerVolume(playerID, 1D);
     }
 
-    public double setVolume(UUID uuid, double value) {
+    public double setPlayerVolume(UUID uuid, double value) {
         volumes.put(uuid, value);
         properties.put(uuid.toString(), String.valueOf(value));
         return value;
     }
 
-    public Map<UUID, Double> getVolumes() {
+    public Map<UUID, Double> getPlayerVolumes() {
         return Collections.unmodifiableMap(volumes);
     }
+
+    public double getCategoryVolume(String category, double def) {
+        Double volume = categoryVolumes.get(category);
+        if (volume == null) {
+            return def;
+        }
+        return volume;
+    }
+
+    public double getCategoryVolume(String category) {
+        return getCategoryVolume(category, 1D);
+    }
+
+    public double setCategoryVolume(String category, double value) {
+        categoryVolumes.put(category, value);
+        properties.put(category, String.valueOf(value));
+        return value;
+    }
+
 }