+++ common/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SelectSpeakerScreen.java
@@ -9,8 +9,8 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import org.jetbrains.annotations.Nullable;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 public class SelectSpeakerScreen extends SelectDeviceScreen {