+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientManager.java
@@ -30,6 +30,7 @@ public class ClientManager {
     @Nullable
     private ClientVoicechat client;
     private final ClientPlayerStateManager playerStateManager;
+    private final ClientCategoryManager categoryManager;
     private final PTTKeyHandler pttKeyHandler;
     private final RenderEvents renderEvents;
     private final KeyEvents keyEvents;
@@ -37,6 +38,7 @@ public class ClientManager {
 
     private ClientManager() {
         playerStateManager = new ClientPlayerStateManager();
+        categoryManager = new ClientCategoryManager();
         pttKeyHandler = new PTTKeyHandler();
         renderEvents = new RenderEvents();
         keyEvents = new KeyEvents();
@@ -166,6 +168,10 @@ public class ClientManager {
 
     public static ClientPlayerStateManager getPlayerStateManager() {
         return instance().playerStateManager;
+    }
+
+    public static ClientCategoryManager getCategoryManager() {
+        return instance().categoryManager;
     }
 
     public static PTTKeyHandler getPttKeyHandler() {