+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/PlayerStateManager.java
@@ -26,6 +26,7 @@ public class PlayerStateManager {
         CommonCompatibilityManager.INSTANCE.onPlayerLoggedOut(this::onPlayerLoggedOut);
         CommonCompatibilityManager.INSTANCE.onServerVoiceChatConnected(this::onPlayerVoicechatConnect);
         CommonCompatibilityManager.INSTANCE.onServerVoiceChatDisconnected(this::onPlayerVoicechatDisconnect);
+        CommonCompatibilityManager.INSTANCE.onPlayerCompatibilityCheckSucceeded(this::onPlayerCompatibilityCheckSucceeded);
 
         CommonCompatibilityManager.INSTANCE.getNetManager().updateStateChannel.setServerListener((server, player, handler, packet) -> {
             PlayerState state = states.get(player.getUUID());
@@ -48,7 +49,7 @@ public class PlayerStateManager {
         voicechatServer.getServer().getPlayerList().getPlayers().forEach(p -> NetManager.sendToClient(p, packet));
     }
 
-    public void onPlayerCompatibilityCheckSucceeded(ServerPlayer player) {
+    private void onPlayerCompatibilityCheckSucceeded(ServerPlayer player) {
         PlayerStatesPacket packet = new PlayerStatesPacket(states);
         NetManager.sendToClient(player, packet);
         Voicechat.logDebug("Sending initial states to {}", player.getDisplayName().getString());