+++ fabric/src/main/java/de/maxhenkel/voicechat/events/ServerVoiceChatEvents.java
@@ -21,4 +21,10 @@ public class ServerVoiceChatEvents {
         }
     });
 
+    public static final Event<Consumer<ServerPlayer>> VOICECHAT_COMPATIBILITY_CHECK_SUCCEEDED = EventFactory.createArrayBacked(Consumer.class, (listeners) -> (player) -> {
+        for (Consumer<ServerPlayer> listener : listeners) {
+            listener.accept(player);
+        }
+    });
+
 }